[Gold V] Title: 옥상 정원 꾸미기, Time: 40 ms, Memory: 6360 KB -BaekjoonHub

+++ 백준/Gold/6198. 옥상 정원 꾸미기/옥상 정원 꾸미기.cc
@@ -9,7 +9,7 @@ using namespace std;
 
 int N;
 vector <int> num;
-map <int, long long> cnt;
+map <int, int> cnt;
 
 int main() {
     ios_base::sync_with_stdio(false);
